refactor: split app to small components
