+++ src/components/AppTitle.vue
@@ -1,0 +1,10 @@
+<script setup>
+import DocumentIcon from "./icons/IconDocument.vue";
+</script>
+
+<template setup>
+  <div class="flex items-center">
+    <DocumentIcon class="mr-2" />
+    <h1 class="text-3xl">Todos App</h1>
+  </div>
+</template>
