+++ src/components/NavBar.vue
@@ -1,3 +1,7 @@
+<script setup>
+import MenuIcon from './icons/IconMenu.vue';
+</script>
+
 <template>
   <nav class="bg-white border-gray-200 dark:bg-gray-900">
     <div
@@ -22,21 +26,7 @@
         aria-expanded="false"
       >
         <span class="sr-only">Open main menu</span>
-        <svg
-          class="w-5 h-5"
-          aria-hidden="true"
-          xmlns="http://www.w3.org/2000/svg"
-          fill="none"
-          viewBox="0 0 17 14"
-        >
-          <path
-            stroke="currentColor"
-            stroke-linecap="round"
-            stroke-linejoin="round"
-            stroke-width="2"
-            d="M1 1h15M1 7h15M1 13h15"
-          />
-        </svg>
+        <MenuIcon />
       </button>
       <div class="hidden w-full md:block md:w-auto" id="navbar-default">
         <ul
