+++ src/components/SearchBar.vue
@@ -1,0 +1,26 @@
+<script setup>
+import SearchIcon from "./icons/IconSearch.vue";
+
+defineProps(["modelValue"]);
+defineEmits(["update:modelValue"]);
+</script>
+
+<template>
+  <form class="mb-5 w-96">
+    <label for="default-search" class="mb-2 text-sm font-medium text-gray-900 sr-only dark:text-white">Search</label>
+    <div class="relative">
+      <div class="absolute inset-y-0 left-0 flex items-center pl-3 pointer-events-none">
+        <SearchIcon />
+      </div>
+      <input
+        type="search"
+        id="default-search"
+        class="block w-full p-4 pl-10 text-sm text-gray-900 border border-gray-300 rounded-lg bg-gray-50 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+        placeholder="Search..."
+        :value="modelValue"
+        @input="$emit('update:modelValue', $event.target.value)"
+        required
+      />
+    </div>
+  </form>
+</template>
